requirements.txt
# FastAPI and server
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# Database
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.30.0
alembic>=1.14.0

# Auth
pyjwt[crypto]>=2.10.0
requests>=2.32.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
google-auth>=2.36.0

# Settings and validation
pydantic[email]>=2.10.0
pydantic-settings>=2.6.0

# Rate limiting
slowapi>=0.1.9

# Utilities
python-multipart>=0.0.17

# Development
black>=24.0.0
ruff>=0.8.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.24.0
httpx>=0.27.0
